refactor: simplify the component layers
