+++ packages/twenty-front/src/modules/workflow/workflow-diagram/components/WorkflowDiagramStepNodeEditableContent.tsx
@@ -1,4 +1,5 @@
 import { WorkflowDiagramStepNodeBase } from '@/workflow/workflow-diagram/components/WorkflowDiagramStepNodeBase';
+import { WorkflowDiagramStepNodeIcon } from '@/workflow/workflow-diagram/components/WorkflowDiagramStepNodeIcon';
 import { WorkflowDiagramStepNodeData } from '@/workflow/workflow-diagram/types/WorkflowDiagram';
 import { WorkflowDiagramNodeVariant } from '@/workflow/workflow-diagram/types/WorkflowDiagramNodeVariant';
 import { FloatingIconButton, IconTrash } from 'twenty-ui';
@@ -16,8 +17,10 @@ export const WorkflowDiagramStepNodeEditableContent = ({
 }) => {
   return (
     <WorkflowDiagramStepNodeBase
-      data={data}
+      name={data.name}
       variant={variant}
+      nodeType={data.nodeType}
+      Icon={<WorkflowDiagramStepNodeIcon data={data} />}
       RightFloatingElement={
         selected ? (
           <FloatingIconButton
