+++ packages/twenty-front/src/modules/workflow/workflow-diagram/components/WorkflowDiagramStepNodeReadonly.tsx
@@ -1,4 +1,5 @@
 import { WorkflowDiagramStepNodeBase } from '@/workflow/workflow-diagram/components/WorkflowDiagramStepNodeBase';
+import { WorkflowDiagramStepNodeIcon } from '@/workflow/workflow-diagram/components/WorkflowDiagramStepNodeIcon';
 import { WorkflowDiagramStepNodeData } from '@/workflow/workflow-diagram/types/WorkflowDiagram';
 
 export const WorkflowDiagramStepNodeReadonly = ({
@@ -6,5 +7,12 @@ export const WorkflowDiagramStepNodeReadonly = ({
 }: {
   data: WorkflowDiagramStepNodeData;
 }) => {
-  return <WorkflowDiagramStepNodeBase variant="default" data={data} />;
+  return (
+    <WorkflowDiagramStepNodeBase
+      name={data.name}
+      variant="default"
+      nodeType={data.nodeType}
+      Icon={<WorkflowDiagramStepNodeIcon data={data} />}
+    />
+  );
 };
